+++ src/Scenario/Transformer/UrlIsTransformer.php
@@ -2,7 +2,6 @@
 
 namespace Automate\Scenario\Transformer;
 
-use Automate\Driver\DriverManager;
 use Facebook\WebDriver\WebDriverExpectedCondition;
 
 class UrlIsTransformer extends AbstractTransformer {