+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.4.0] - Coming
+## [0.4.1] - 28/02/2021
+
+### Fixed
+
+- Scenario reset the variable scope each time you call the constructor
+
+## [0.4.0] - 27/02/2021
 
 ### Added